requirements.txt
flask
openai
python-dotenv
whisper
yt-dlp
google-cloud-texttospeech
